🚑 metadata 적용을 위한 클라이언트/서버 컴포넌트 분리
